requirements.txt
python-dotenv
langchain
langchain-mistralai
langchain-core
requests
pandas